+++ packages/sdk/shopify-oxygen/src/global.d.ts
@@ -1,0 +1,7 @@
+/**
+ * Shopify Oxygen Workers Types
+ * @see https://shopify.dev/docs/storefronts/headless/hydrogen/deployments/oxygen-runtime
+ */
+import '@shopify/oxygen-workers-types';
+
+export {};