+++ packages/sdk/shopify-oxygen/src/platform/OxygenCrypto/index.ts
@@ -1,0 +1,19 @@
+import type { Crypto, Hasher, Hmac } from '@launchdarkly/js-server-sdk-common';
+
+import CryptoJSHasher from './cryptoJSHasher';
+import CryptoJSHmac from './cryptoJSHmac';
+import { SupportedHashAlgorithm } from './types';
+
+export default class EdgeCrypto implements Crypto {
+  createHash(algorithm: SupportedHashAlgorithm): Hasher {
+    return new CryptoJSHasher(algorithm);
+  }
+
+  createHmac(algorithm: SupportedHashAlgorithm, key: string): Hmac {
+    return new CryptoJSHmac(algorithm, key);
+  }
+
+  randomUUID(): string {
+    return crypto.randomUUID();
+  }
+}