+++ packages/sdk/shopify-oxygen/src/platform/index.ts
@@ -1,0 +1,19 @@
+import type { Info } from '@launchdarkly/js-server-sdk-common';
+import { platform } from '@launchdarkly/js-server-sdk-common';
+
+import { OxygenCacheOptions } from '../utils/validateOptions';
+import OxygenCrypto from './OxygenCrypto';
+import OxygenRequests from './OxygenRequests';
+
+export default class OxygenPlatform implements platform.Platform {
+  info: Info;
+
+  crypto: platform.Crypto = new OxygenCrypto();
+
+  requests: platform.Requests;
+
+  constructor(info: Info, cacheOptions: OxygenCacheOptions = {}) {
+    this.info = info;
+    this.requests = new OxygenRequests(cacheOptions);
+  }
+}